backend-potato/requirements.txt
flask==2.3.3
tensorflow==2.10.0
Pillow==9.5.0
numpy==1.23.5
